requirements.txt
autopep8==2.0.2
cachetools==5.3.0
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.1.0
clowncar==0.2.0
contourpy==1.0.7
cycler==0.11.0
DateTime==5.1
fonttools==4.39.3
google-auth==2.17.3
google-auth-oauthlib==1.0.0
groupy==0.5.6
GroupyAPI==0.10.3
gspread==5.8.0
idna==2.7
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.2
oauthlib==3.2.2
packaging==23.1
pandas==2.0.0
Pillow==9.5.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.20.1
requests-oauthlib==1.3.1
rsa==4.9
seaborn==0.12.2
six==1.16.0
tomli==2.0.1
tornado==6.2
typing==3.7.4.3
tzdata==2023.3
urllib3==1.24.3
zope.interface==6.0
